+++ src/main/java/com/danielbyrne/daftsearch/services/PropertyForSaleService.java
@@ -12,7 +12,7 @@ public interface PropertyForSaleService {
 
     Set<PropertyForSaleDTO> getAllProperties();
 
-    Set<PropertyForSaleDTO> filterPropertiesByDaftAttributes(int maxPrice, int minBed, County[] counties);
+    Set<PropertyForSaleDTO> filterPropertiesByDaftAttributes(float maxPrice, int minBed, County[] counties);
 
     Set<PropertyForSaleDTO> filterPropertiesByGoogle(Set<PropertyForSaleDTO> preFilteredDTOs, String origin,
                                                      ModeOfTransport modeOfTransport, int distance, int duration)
